Rename opentelemetry-instrumentation-api module to v1.14 and add v1_14 subpackage
Renames the directory to introduce a base version (`opentelemetry-instrumentation-api-1.14`,
matching the muzzle floor of `1.14.0-alpha`) and moves the javaagent classes into a
`v1_14` subpackage, bringing this module in line with the convention used by the other
self-instrumentation modules (`opentelemetry-extension-annotations-1.0`,
`opentelemetry-instrumentation-annotations-1.16`).
