+++ .fossa.yml
@@ -222,9 +222,6 @@ targets:
       target: ':instrumentation:opentelemetry-instrumentation-annotations-1.16:javaagent'
     - type: gradle
       path: ./
-      target: ':instrumentation:opentelemetry-instrumentation-api:javaagent'
-    - type: gradle
-      path: ./
       target: ':instrumentation:oracle-ucp-11.2:javaagent'
     - type: gradle
       path: ./
@@ -847,6 +844,9 @@ targets:
     - type: gradle
       path: ./
       target: ':instrumentation:opentelemetry-api:opentelemetry-api-1.59:javaagent'
+    - type: gradle
+      path: ./
+      target: ':instrumentation:opentelemetry-instrumentation-api:opentelemetry-instrumentation-api-1.14:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:pekko:pekko-actor-1.0:javaagent'
