--- instrumentation/opentelemetry-instrumentation-api/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/instrumentationapi/InstrumentationApiInstrumentationModule.java
@@ -1,33 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.instrumentationapi;
-
-import static java.util.Arrays.asList;
-
-import com.google.auto.service.AutoService;
-import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.internal.ExperimentalInstrumentationModule;
-import java.util.List;
-
-@AutoService(InstrumentationModule.class)
-public class InstrumentationApiInstrumentationModule extends InstrumentationModule
-    implements ExperimentalInstrumentationModule {
-
-  public InstrumentationApiInstrumentationModule() {
-    super("opentelemetry-instrumentation-api");
-  }
-
-  @Override
-  public List<TypeInstrumentation> typeInstrumentations() {
-    return asList(new HttpRouteStateInstrumentation(), new SpanKeyInstrumentation());
-  }
-
-  @Override
-  public String getModuleGroup() {
-    return "opentelemetry-api-bridge";
-  }
-}
