+++ instrumentation/opentelemetry-instrumentation-api/opentelemetry-instrumentation-api-1.14/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/instrumentationapi/v1_14/InstrumentationApiInstrumentationModule.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.instrumentationapi.v1_14;
+
+import static java.util.Arrays.asList;
+
+import com.google.auto.service.AutoService;
+import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
+import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
+import io.opentelemetry.javaagent.extension.instrumentation.internal.ExperimentalInstrumentationModule;
+import java.util.List;
+
+@AutoService(InstrumentationModule.class)
+public class InstrumentationApiInstrumentationModule extends InstrumentationModule
+    implements ExperimentalInstrumentationModule {
+
+  public InstrumentationApiInstrumentationModule() {
+    super("opentelemetry-instrumentation-api", "opentelemetry-instrumentation-api-1.14");
+  }
+
+  @Override
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return asList(new HttpRouteStateInstrumentation(), new SpanKeyInstrumentation());
+  }
+
+  @Override
+  public String getModuleGroup() {
+    return "opentelemetry-api-bridge";
+  }
+}
