+++ settings.gradle.kts
@@ -532,8 +532,8 @@ include(":instrumentation:opentelemetry-api:opentelemetry-api-1.61:testing")
 include(":instrumentation:opentelemetry-extension-annotations-1.0:javaagent")
 include(":instrumentation:opentelemetry-extension-kotlin-1.0:javaagent")
 include(":instrumentation:opentelemetry-instrumentation-annotations-1.16:javaagent")
-include(":instrumentation:opentelemetry-instrumentation-api:javaagent")
-include(":instrumentation:opentelemetry-instrumentation-api:testing")
+include(":instrumentation:opentelemetry-instrumentation-api:opentelemetry-instrumentation-api-1.14:javaagent")
+include(":instrumentation:opentelemetry-instrumentation-api:opentelemetry-instrumentation-api-1.14:testing")
 include(":instrumentation:oracle-ucp-11.2:javaagent")
 include(":instrumentation:oracle-ucp-11.2:library")
 include(":instrumentation:oracle-ucp-11.2:testing")
